+++ app/src/main/AndroidManifest.xml
@@ -18,11 +18,7 @@
             <meta-data
                 android:name="android.app.default_searchable"
                 android:value=".Activity.SearchResultsActivity" />
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
 
         </activity>
         <!-- Sub Page activity -->
@@ -43,6 +39,14 @@
         </activity>
 
         <activity android:name=".Activity.NewsTabActivity">
+        </activity>
+
+        <activity android:name=".Activity.SplashActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
 
     </application>
