+++ app/src/main/java/uz/chamber/maroqand/Activity/Main.java
@@ -60,7 +60,6 @@ public class Main extends AppCompatActivity
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.main);
-        changeLanguage(AppConfig.getLanguage());
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
@@ -72,6 +71,7 @@ public class Main extends AppCompatActivity
         toggle.syncState();
 
         NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
+        navigationView.invalidate();
         navigationView.setNavigationItemSelectedListener(this);
 
         viewPager = (ViewPager) findViewById(R.id.vp_mainViewPager);
@@ -210,14 +210,7 @@ public class Main extends AppCompatActivity
 
     }
 
-    public void changeLanguage(String languageToLoad) {
-        //for language change
-        Locale locale = new Locale(languageToLoad);
-        Locale.setDefault(locale);
-        Configuration config = new Configuration();
-        config.locale = locale;
-        getResources().updateConfiguration(config, getResources().getDisplayMetrics());
-    }
+
 
 
     @Override
